Implement Layer 0 non-factorized optimizations to Gradient computation
Normally, when using N threads, we compute N gradients for all weights while iterating over the batch and computing the backprop. Now, we using a single Gradient for the main input weights [0..40960], while still using N Gradients for the factorized weights.

We lock each layer of the single Gradient, and update. Local testing appears that there is little diffrence in speed and result when not using locks on 16 threads. This saves us from a massive amount of effort in zero'ing Gradients, as well as avoiding having to sum up N Gradients.

This change gained 27% speed on a Ryzen 3700x using 16-threads with a batchsize of 16384. Smaller batchsizes will see even larger gains. Now, for the first time, the use of hyperthreads is preferable to cores.